requirements.txt
appnope
astroid
atomicwrites
attrs
backcall
certifi
chardet
decorator
idna
ipdb
ipython
ipython-genutils
isort
jedi
lazy-object-proxy
lxml
mccabe
mock
more-itertools
parso
pbr
pexpect
pickleshare
pluggy
prompt-toolkit
psycopg2-binary
ptyprocess
py
pydantic
Pygments
pylint
pytest
PyYAML
requests
requests-mock
simplejson
six
traitlets
typed-ast
urllib3
wcwidth
wrapt
